[BUG Fixed] fix dist_op building when using PyTorch 2.0 (#59)
**Description**
In PyTorch 2.0, when building `msamp_dist_op`, it raises the following
error:
```
    /root/ms-amp/msamp/operators/dist_op/dist.cpp:209:38: error: no matching function for call to ‘pybind11::module_::def(const char [7], <unresolved overloaded function type>, const char [7])’
                                  209 |     m.def("reduce", &reduce, "Reduce");
                                      |                                      ^   
```

The cause is that there is a function called `reduce` in torch
namespace. The function name is the same as that of our custom function.

In this PR, I rename the two functions `reduce` and `all_reduce`. Fix
#42

**Major Revision**
- rename two functions, i.e. `reduce` and `all_reduce` in
`dist_op/dist.cpp`.
- remove `using namespace torch/std;` in dist_op
- link system nccl by adding `libraries=['nccl']` in setup.py