Add cifar10 example using TE+DeepSpeed-Zero+MS-AMP (#143)
**Description**
Add cifar10 example using TE+DeepSpeed-Zero+MS-AMP
There is a 1P customer who only uses DeepSpeed ZeRO and the model is
transformer based. Purely replacing Linear with FP8Linear will not get
performance gain. So we suggest them to use TransformerEngine for
speeding up and MS-AMP for saving memory. This is a good reference to
them.

**Major Revision**
- Add example
- Add document
- Fix a bug in mnist_ddp example